\name{BEDASSLE-internal}
\alias{a0_gibbs_rate}
\alias{BB_Likelihood_counts}
\alias{BB_Prior_prob_phi}
\alias{BB_Update_mu}
\alias{BB_Update_phi}
\alias{BB_Update_thetas}
\alias{identify_invariant_loci}
\alias{Initialize.params}
\alias{Likelihood_counts}
\alias{Likelihood_thetas}
\alias{load_MCMC_output}
\alias{load_posterior_predictive_samples}
\alias{Prior_prob_alpha0}
\alias{Prior_prob_alpha2}
\alias{Prior_prob_alphaD}
\alias{Prior_prob_alphaE}
\alias{Prior_prob_beta}
\alias{Prior_prob_mu}
\alias{Shift}
\alias{simulate_allele_count_data}
\alias{transform_frequencies}
\alias{Update_a0}
\alias{Update_a2}
\alias{Update_aD}
\alias{Update_aE}
\alias{Update_beta}
\alias{Update_mu}
\alias{Update_thetas}
\title{Internal BEDASSLE Functions}
\description{
 Internal BEDASSLE functions
}
\details{
These functions are called by other functions (mostly
\command{MCMC} and \command{MCMC_BB}), and will not 
be called directly by the user.
}