% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.data.R
\name{freqs2xval}
\alias{freqs2xval}
\title{Create replicate data partitions for k-fold cross-validation}
\usage{
freqs2xval(freqs, n.replicates = 10, n.partitions = 10, prefix)
}
\arguments{
\item{freqs}{A \code{matrix} of allele frequencies with 
one column per locus and one row per sample.
    Missing data should be indicated with \code{NA}.}

\item{n.replicates}{An integer giving the number (n) of cross-validation 
replicates to be performed. Default is 10.}

\item{n.partitions}{An integer giving the number (k) of partitions (folds)
 into which the dataset should be divided for k-fold cross-validation. 
Default is 10.}

\item{prefix}{A character \code{vector} giving the prefix to be attached 
to the output file of replicated data partitions. An underscore is 
automatically added between the prefix and the file names.}
}
\value{
This function generates and saves a list of length \code{n.replicates}, 
	each element of which is a list of length \code{n.partitions}. text files, each containing the 
	pairwise pi matrix calculated from its partition of the allele frequency data 
	matrix. These files can then be used for running a k-fold cross-validation 
	analysis using \code{\link{x.validation}}.
}
\description{
\code{freqs2xval} creates a list of data partitions to be 
used in an n-replicate k-fold cross-validation analysis.
divides the allele frequency dataset into partitions, and, within each, calculates 
pairwise pi between samples for use in k-fold cross-validation.
}
\details{
This function takes an allele frequency data matrix
for a sample of diploid individuals, and, for each cross-validation 
replicate, divides it randomly into \emph{k} partitions and 
calculates pairwise pi between all samples within each partition.
These replicate partitions can then be used in a cross-validation 
analysis.

For each of \code{n.replicates}, this function divides an 
	allele frequency dataset randomly into \emph{k} equal partitions 
	and, within each, calculates pairwise pi (the proportion of 
	sites at which each pair of samples differs, out of the total 
	number of loci in the dataset) between a set of diploid 
	individuals. For details on how pairwise pi is calculated, see 
	\code{\link{freqs2pairwisePi}}. Pairwise pi within each data partition 
	is saved as a text file, and these can then be used to run a cross-validation 
	analysis with \code{\link{x.validation}}. 
	
		In each replicate, the entire dataset is divided evenly into the \emph{k}, 
	the number of partitions specified in \code{n.partitions}. If the total number of 
	loci \emph{L} cannot be evenly divided into \emph{k} partitions, the remainder of 
	\emph{L}/\emph{k} loci are dropped from the dataset. Within each partition, 
	there must be more loci than there are samples.
}
