% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.output.R
\name{make.all.bedassle.plots}
\alias{make.all.bedassle.plots}
\title{Make output plots}
\usage{
make.all.bedassle.plots(results.files, data.block.file, prefix,
  chain.cols = NULL)
}
\arguments{
\item{results.files}{A \code{vector} of the filenames (in quotes, 
with the full file path) of the posterior results files 
output by the different chains of a \code{BEDASSLE} run. 
Can also be a single filename if user wants to visualize 
only a single run.}

\item{data.block.file}{The filename (in quotes, with the full file 
path) of the data.block R object (.Robj) file output by a 
\code{BEDASSLE} run.}

\item{prefix}{A character vector to be prepended to all figures.}

\item{chain.cols}{A \code{vector} of colors to be used in plotting 
results from different MCMC chains. There should be one 
color specified for each chain. If \code{NULL}, the plots 
will use the required set or subset of 12 pre-specified 
colors. If there are more than 12 chains, users must supply 
their own colors.}
}
\value{
This function has only invisible return values.
}
\description{
\code{make.all.bedassle.plots} makes figures from the output from a 
	BEDASSLE analysis.
}
\details{
This function takes the file output from a BEDASSLE analysis and 
generates a number of plots for visualizing results and 
diagnosing MCMC performance.

This function produces a variety of plots that can be 
useful for visualizing results or diagnosing MCMC performance. 
 The plots made are by no means exhaustive, and users are 
	encouraged to make further plots, or customize these plots as they 
see fit. The plots generated (as .pdf files) are:
\itemize{
	\item model.fit.CIs - A plot of the sample allelic covariance 
		shown with the 95\% credible interval of the parametric 
		covariance for each entry in the matrix. Only generated 
		if either the \code{geoDist} or \code{envDist} arguments 
		in the \code{run.bedassle} function call are specified. One 
		plot is produced for each chain.
	\item Trace plots - Plots of parameter values over the MCMC.
	\itemize{
		\item lpd - A plot of the log posterior probability over the MCMC.
		\item nuggets - A plot of estimates of the nugget parameters 
			over the MCMC.
		\item gamma - A plot of estimates of the gamma parameter 
			over the MCMC.
		\item alpha parameters - Plots of estimates of the 
			various parameters (all or some of {alpha0,alphaD,alphaE,alpha2}, 
			depending on the model specified) over the MCMC.
	}
}
}
