% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comparison.R
\name{makePartitions}
\alias{makePartitions}
\title{Make the partitions used to run a BEDASSLE cross-validation analysis}
\usage{
makePartitions(prefix, nReplicates = 10, nPartitions = 5, N)
}
\arguments{
\item{prefix}{A \code{character} object giving the prefix to be attached 
to the output data partitions object to be used in the k-fold 
cross-validation procedure.}

\item{nReplicates}{An \code{integer} giving the number of cross-validation
replicates to be run. The default value is 10.}

\item{nPartitions}{An \code{integer} giving the number of data folds 
within each run. The default value is 5.}

\item{N}{An \code{integer} giving the number of samples in the dataset.
This should have the same value as \code{nrow(geoDist)}.}
}
\value{
This function generates and saves a nested list R object (".Robj" file) 
	that can be used to run a cross-validation analysis. This R object is a list 
	of length \code{nReplicates}, each element of which is a list of length 
	\code{nPartitions}. Each of those \code{nPartitions} element 
	is a vector of indices (between 1 and \code{N}) of the random subset 
	of individuals assigned to that cross-validation partition. 
	Each of the \code{N} individuals appears in exactly one partition per replicate.
}
\description{
\code{makePartitions} makes the partitions used to run a BEDASSLE cross-validation analysis
}
\details{
This function initiates a k-fold cross-validation analysis 
to determine the statistical support for the specified model.

The number of samples \code{N} must be divisible by the specified 
	number of partitions (\code{nPartitions}).
}
