% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.bedassle.R
\name{run.bedassle}
\alias{run.bedassle}
\title{Run a BEDASSLE analysis.}
\usage{
run.bedassle(
  genDist,
  geoDist = NULL,
  envDist = NULL,
  nLoci,
  prefix,
  n.chains = 4,
  n.iter = 2000,
  make.figs = TRUE,
  save.stanfit = TRUE,
  ...
)
}
\arguments{
\item{genDist}{A \code{matrix} of pairwise pi measured between 
all pairs of samples.}

\item{geoDist}{A \code{matrix} of pairwise geographic distances 
measured between all pairs of samples. A value of 
\code{NULL} runs a model without geographic distance 
as a predictor of genetic differentiation.}

\item{envDist}{A \code{matrix} of pairwise environmental distances 
measured between all pairs of samples. If there 
are multiple environmental distance measures, this 
argument should be a \code{list} of distance matrices. 
 A value of \code{NULL} runs a model without geographic 
distance as a predictor of genetic differentiation.}

\item{nLoci}{The number of loci used in the calculation of the pairwise 
pi \code{matrix} specified in the \code{genDist} argument.}

\item{prefix}{A character \code{vector} giving the prefix to be attached 
to all output files. An underscore is automatically 
added between the prefix and the file names.}

\item{n.chains}{An integer indicating the number of MCMC chains to be run 
in the analysis. Default is 4.}

\item{n.iter}{An \code{integer} giving the number of iterations each MCMC 
chain is run. Default is 2e3.  If the number of iterations 
is greater than 500, the MCMC is thinned so that the number 
of retained iterations is 500 (before burn-in).}

\item{make.figs}{A \code{logical} value indicating whether to automatically 
make figures once the analysis is complete. Default is 
\code{TRUE}.}

\item{save.stanfit}{A \code{logical} value indicating whether to automatically 
save the full \code{stanfit} output once the analysis is
complete. Default is \code{TRUE}.}

\item{...}{Further options to be passed to rstan::sampling (e.g., adapt_delta).}
}
\value{
This function writes all output to three tab-delimited text files 
		for each chain run (specified with \code{n.chains}). The output files 
		associated with each chain have "chain_X" for the Xth chain appended 
		to the file name, and begin with the prefix specified with the 
		\code{prefix} argument. The three output files associated with each 
		chain are described below:
		\itemize{
			\item \code{posterior} contains the posterior probability and 
					parameter estimates over the sampled posterior distribution
					of the MCMC. Each entry described below has its own named 
					column in the output text file. Columns are separated by tabs.
				\itemize{
					\item \code{lpd} log posterior density over the retained 
							MCMC iterations.
					\item \code{alpha0} posterior draws for alpha0 parameter.
					\item \code{alphaD} posterior draws for alphaD parameter.
					\item \code{alphaE} posterior draws for alphaE parameter(s).
							Named "alphaE_E" for the Eth environmental distance 
							matrix specified.
					\item \code{alpha2} posterior draws for alpha2 parameter.
					\item \code{nuggets} posterior draws for nugget parameters. 
							Named "nugget_N" for the Nth sample in the dataset.
				}
			\item \code{MAP} contains point estimates of the parameters listed in 
							the \code{posterior} file described above. Values are 
							indexed at the MCMC iteration with the greatest 
							posterior probability.
			\item \code{parCov} contains the covariance matrix parameterized by the 
							MAP parameter point estimates.
		}
}
\description{
\code{run.bedassle} runs a BEDASSLE analysis
}
\details{
This function runs an analysis that 
estimates the relative contributions of geographic 
and environmental/ecological distances to patterns of genetic 
differentiation between samples.

This function acts as a wrapper around a STAN model block determined 
		by the user-specified model (e.g., just geographic distance, or geographic distance 
		plus 2 environmental/ecological distance variables).
		User-specified data are checked for appropriate format and consistent dimensions,
		then formatted into a \code{data.block},
		which is then passed to the STAN model block.
		The model output is written to three tab-delimited text files described above.
		The parameter values are rounded to 4 decimal places
		The full \code{stanfit} model output is also saved if \code{save.stanfit=TRUE}.
		If \code{make.figs=TRUE}, running \code{run.bedassle} will also generate figures
		depicting different aspects of model output; these are detailed in the function 
		\code{make.all.bedassle.plots} in this package.
}
