% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comparison.R
\name{compare.model.xvals}
\alias{compare.model.xvals}
\title{Compare the cross-validation output of different models}
\usage{
compare.model.xvals(xval.files, mod.order, mod.names, mod.cols = NULL)
}
\arguments{
\item{xval.files}{A \code{character} vector giving the filenames (in quotes, 
with the full file path) to the output files of a \code{\link{xValidation}} 
analysis.}

\item{mod.order}{An \code{integer} vector giving the ordering of the model hypotheses 
to be compared. Lower numbers indicate less complex or more parsimonious models.
No ties.}

\item{mod.names}{An \code{character} vector giving the names associated with the models 
run in the different cross-validation analyses.}

\item{mod.cols}{A vector containing the colors to be used in plotting the output of 
the different cross-validation analyses. Default is black.}
}
\value{
This function generates a figure comparing the cross-validation predictive accuracy 
	of different models used to analyze the same data partitions. The "best" model among the 
	set being compared is indicated with a golden arrow.  The significance of the pairwise 
	comparisons is indicated using letter groupings, as in a Tukey post-hoc test. The 
	function also returns the name of the "best" model.
}
\description{
\code{compare.model.xvals} compares the outputs of different BEDASSLE cross-validation analyses
}
\details{
This function compares outputs of different BEDASSLE cross-validation analyses 
to determine the relative statistical support for each model and identify the best model.

This function compares the cross-validation predictive accuracy of different models 
	applied to the same data partitions (generated using \code{\link{makePartitions}}). 
	Going in the order of model complexity specified by \code{mod.order}, starting from the 
	least complex model, this function compares the performance of a model across all 
	cross-validation data partitions against that of the next most complex model. The model  
	that is determined to be "best" is the least complicated model whose performance is 
	statistically indistinguishable from adjacent model (ie, the model that is one rung 
	higher in complexity, as specified by \code{mod.order}.
}
