% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.comparison.R
\name{makePartitions}
\alias{makePartitions}
\title{Run a BEDASSLE cross-validation analysis}
\usage{
makePartitions(prefix, nReplicates = 10, nPartitions = 5, N, relMat, nLoci)
}
\arguments{
\item{prefix}{A character \code{vector} giving the prefix to be attached 
to the output data partitions object to be used in the k-fold 
cross-validation procedure.}

\item{nReplicates}{An \code{integer} giving the number of cross-validation
replicates to be run. The default value is 10.}

\item{nPartitions}{An \code{integer} giving the number of data folds 
within each run. The default value is 5.}

\item{N}{An \code{integer} giving the number of samples in the dataset.
This should have the same value as \code{nrow(geoDist)}.}

\item{relMat}{A \code{matrix} of pairwise genetic distances 
measured between all pairs of samples.}

\item{nLoci}{The total number of loci in the dataset.}
}
\value{
This function generates and saves a nested list R object (".Robj" file) 
	that can be used to run a cross-validation analysis. This R object is a list 
	of length \code{nReplicates}, each element of which is a list of length 
	\code{nPartitions}. The contents of each of those \code{nPartitions} elements 
	is given below:
	\itemize{
		\item trainBlock
			\itemize{
				\item N the number of samples in the training data partition
				\item L the number of loci in the dataset
				\item obsCov the pairwise relatedness matrix between 
					all samples in the training data partition
			}
		\item testBlock
			\itemize{
				\item N the number of samples in the testing data partition
				\item L the number of loci in the dataset
				\item obsCov the pairwise relatedness matrix between 
					all samples in the testing data partition
			}
		\item inTest the indices of the samples in the testing data partition
	}
}
\description{
\code{x.validation} runs a BEDASSLE cross-validation analysis
}
\details{
This function initiates a k-fold cross-validation analysis 
to determine the statistical support for the specified model.
}
